Add Logout action to AuthController and update link
Added a new Logout action method to the AuthController class in the
CodeNest.UI.Controllers namespace. This method clears the user's
session and redirects them to the Login page.

Updated the Default.cshtml file to change the "Log Out" link. The link
now uses ASP.NET Core tag helpers to call the Logout action in the
AuthController instead of using a JavaScript void function.